DataEng/backend/requirements.txt
# Backend API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart

# Template Engine
jinja2>=3.1.2
pyyaml>=6.0.1
networkx>=3.1

# Python-Native TUI
textual>=0.50.0
rich>=13.7.0
colorama>=0.4.6

# Streamlit UI (Optional)
streamlit>=1.30.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
